Return all Handles by Value instead of by ref
Returning Handles by reference will cause many issues
as the underlying object might be destroyed by python
which leads to an invalid reference itself. By returning
it by value we enforce, that a new Handle is created
that cannot be destroyed be the creator object.

This commit returns the Handles by value by a simple
change in the swig interface files. This change is
required if we want to get rid of the garbage collector.

See discussion #214 in tpaviot/pythonocc-core.