Fixes a memory management error
This commit tries to fix the issue tpaviot/pythonocc-core#292.
When we got a handle from an API function and called
GetObject() the lifetime of the object was determined
only by the handle. This lead to crashes, when only a
reference to the object was stored. The commit registers
the handle with its object to prevent premature destruction.